third_party/blink/web_tests/platform/linux-chrome/external/wpt/mathml/presentation-markup/tables/table-default-styles-001-expected.txt

This is a testharness.js-based test.
[FAIL] Default display value on mtable
  assert_equals: expected "inline-table" but got "table"
[FAIL] Default text-align value on mtd
  assert_equals: expected "center" but got "-webkit-center"
[FAIL] Default padding-top value on mtd
  assert_equals: expected "10px" but got "6px"
[FAIL] Default padding-bottom value on mtd
  assert_equals: expected "10px" but got "6px"
[FAIL] Bounding box is the same as mrow + mtd default padding
  assert_equals: expected 30 but got 22
Harness: the test ran to completion.
